refactor(template): cleanup
